Fixes biome "unlocking" on purchase.
The bug was in "buy" command, especially with a custom gui's where players get biome with `buy` command. In situations, where biome was added after player re-logged, it was not marked as "unlocked", because validation happens only on login and build-in gui opening. This change will add a new validation that happens inside 'buy' command, and will unlock biome if it is possible.

Fixes #105